+++ Brandingdong/Brandingdong/API/Service.swift
@@ -7,9 +7,80 @@
 //
 
 import UIKit
+import Alamofire
 
 struct Service {
   
+  static func getBannerImages(completion: @escaping (Bool) -> Void ) {
+    
+    let bannerUrl = "http://52.78.75.94/events/"
+    AF.request(bannerUrl, method: .get).responseJSON { response in
+      guard let jsonData = response.data else { return }
+      do {
+        let bannerImages = try JSONDecoder().decode([HomeInfoCodAble.Events].self, from: jsonData)
+        for index in 0..<bannerImages.count {
+          HomeInfoDatas.bannerImages.append(bannerImages[index].images)
+        }
+        DispatchQueue.main.async {
+          completion(true)
+        }
+      } catch {
+        print ("error : ", error.localizedDescription)
+      }
+    }
+  }
+  
+  static func getProductList(completion: @escaping (Bool) -> Void) {
+    let productUrl = "http://52.78.75.94/products/detail"
+    
+    AF.request(productUrl, method: .get).responseJSON { response in
+      guard let jsonData = response.data else { return }
+      do {
+        let productDatas = try JSONDecoder().decode(HomeInfoCodAble.ProductList.self, from: jsonData)
+        
+        for index in 0..<productDatas.results.count {
+          
+          let name = productDatas.results[index].name
+          let price = productDatas.results[index].price
+          let mainImages = productDatas.results[index].main_img[0].image
+          let brandName = productDatas.results[index].brand.name
+          let brandImages = productDatas.results[index].brand.brand_img
+          let brandIntro = productDatas.results[index].brand.intro
+          
+          HomeInfoDatas.names.append(name)
+          HomeInfoDatas.price.append(price)
+          HomeInfoDatas.images.append(mainImages)
+          HomeInfoDatas.brandNames.append(brandName)
+          
+          if !(HomeInfoDatas.productNameAndBrandImageIntro.keys.contains(name)) {
+            HomeInfoDatas.productNameAndBrandImageIntro[name] = [brandIntro : brandImages]
+          } else {
+            HomeInfoDatas.productNameAndBrandImageIntro[name]?.updateValue(brandImages, forKey: brandIntro)
+          }
+          
+          if !(HomeInfoDatas.productNameAndBrandNamePrice.keys.contains(name)) {
+            HomeInfoDatas.productNameAndBrandNamePrice[name] = [brandName : price]
+          } else {
+            HomeInfoDatas.productNameAndBrandNamePrice[name]?.updateValue(price, forKey: brandName)
+          }
+          
+          for imageIndex in 0..<productDatas.results[index].main_img.count {
+            if !(HomeInfoDatas.productNameAndImages.keys.contains(name)) {
+              HomeInfoDatas.productNameAndImages[name] = [imageIndex : productDatas.results[index].main_img[imageIndex].image]
+            } else {
+              HomeInfoDatas.productNameAndImages[name]?.updateValue(productDatas.results[index].main_img[imageIndex].image, forKey: imageIndex)
+            }
+          }
+          DispatchQueue.main.async {
+            completion(true)
+          }
+        }
+      } catch {
+        print ("failed to convert error : ", error.localizedDescription)
+      }
+    }
+  }
+
   static func signUpUser(username: String, email: String, password1: String, password2: String, phonenumber: String) {
     let signUpUrl = "http://52.78.75.94/auth/signup/"
     guard let url = URL(string: signUpUrl) else  { return print ("can't not create url")}
@@ -68,5 +139,4 @@ struct Service {
     }
     task.resume()
   }
-  
 }
