+++ Brandingdong/Brandingdong/Controller/Brandingdong/Controller/Home/HomeLaunchScreenViewController.swift
@@ -27,34 +27,6 @@ class HomeLaunchScreenViewController: UIViewController {
     }
   }
   
-  // MARK: - Json Codable
-  
-  struct ProductList: Codable {
-    var results: [Results]
-    
-    struct Results: Codable {
-      var id: Int
-      var name: String
-      var price: Int
-      var main_img: [Images]
-      var brand: Brand
-    }
-    
-    struct Images : Codable {
-      var image : String
-    }
-    
-    struct Brand: Codable {
-      var name: String
-      var brand_img: String
-    }
-  }
-  
-  struct Events : Codable {
-    var images : String
-  }
-  
-  
   // MARK: - LifeCycle
   
   override func viewDidLoad() {
@@ -62,8 +34,7 @@ class HomeLaunchScreenViewController: UIViewController {
     setUI()
     setConstraints()
     setIndicatorView()
-    getProductList()
-    getBannerImages()
+    HomeInfoGetData()
   }
   
   override func viewWillAppear(_ animated: Bool) {
@@ -103,68 +74,16 @@ class HomeLaunchScreenViewController: UIViewController {
     indicatorView.startAnimating()
   }
   
-  func getProductList() {
-    let productUrl = "http://52.78.75.94/products/detail"
+  private func HomeInfoGetData() {
     
-    AF.request(productUrl, method: .get).responseJSON { response in
-      guard let jsonData = response.data else { return }
-      do {
-        let productDatas = try JSONDecoder().decode(ProductList.self, from: jsonData)
-        
-        for index in 0..<productDatas.results.count {
-          
-          let name = productDatas.results[index].name
-          let price = productDatas.results[index].price
-          let mainImages = productDatas.results[index].main_img[0].image
-          let brandName = productDatas.results[index].brand.name
-          let brandImages = productDatas.results[index].brand.brand_img
-          
-          HomeInfoDatas.names.append(name)
-          HomeInfoDatas.price.append(price)
-          HomeInfoDatas.images.append(mainImages)
-          HomeInfoDatas.brandNames.append(brandName)
-          
-          if !(HomeInfoDatas.productNameAndBrandNamePrice.keys.contains(name)) {
-            HomeInfoDatas.productNameAndBrandImage[name] = brandImages
-          } else {
-            HomeInfoDatas.productNameAndBrandImage.updateValue(brandImages, forKey: name)
-          }
-          
-          if !(HomeInfoDatas.productNameAndBrandNamePrice.keys.contains(name)) {
-            HomeInfoDatas.productNameAndBrandNamePrice[name] = [brandName : price]
-          } else {
-            HomeInfoDatas.productNameAndBrandNamePrice[name]?.updateValue(price, forKey: brandName)
-          }
-          
-          for imageIndex in 0..<productDatas.results[index].main_img.count {
-            if !(HomeInfoDatas.productNameAndImages.keys.contains(name)) {
-              HomeInfoDatas.productNameAndImages[name] = [imageIndex : productDatas.results[index].main_img[imageIndex].image]
-            } else {
-              HomeInfoDatas.productNameAndImages[name]?.updateValue(productDatas.results[index].main_img[imageIndex].image, forKey: imageIndex)
-            }
-          }
-          self.firstToggle = true
-          
-        }
-      } catch {
-        print ("failed to convert error : ", error.localizedDescription)
-      }
+    Service.getProductList { (isSuccess) in
+      guard isSuccess else { return }
+      self.firstToggle = true
     }
-  }
-  
-  private func getBannerImages() {
-    let bannerUrl = "http://52.78.75.94/events/"
-    AF.request(bannerUrl, method: .get).responseJSON { response in
-      guard let jsonData = response.data else { return }
-      do {
-        let bannerImages = try JSONDecoder().decode([Events].self, from: jsonData)
-        for index in 0..<bannerImages.count {
-          HomeInfoDatas.bannerImages.append(bannerImages[index].images)
-          self.secondToggle = true
-        }
-      } catch {
-        print ("error : ", error.localizedDescription)
-      }
+    
+    Service.getBannerImages { (isSuccess) in
+      guard isSuccess else { return }
+      self.secondToggle = true
     }
   }
   
@@ -181,66 +100,3 @@ class HomeLaunchScreenViewController: UIViewController {
   }
 }
 
-// MARK: - Test Code
-
-/*
- private func getProductListUrlSession() {
- 
- let productUrl = "http://52.78.75.94/products/detail"
- guard let url = URL(string: productUrl) else { return }
- let task = URLSession.shared.dataTask(with: url) { (data, response, error) in
- guard error == nil else { return print ("error : ", error!.localizedDescription)}
- let responsea = response as? HTTPURLResponse
- print (responsea!.statusCode)
- guard let response = response as? HTTPURLResponse,
- (200..<406).contains(response.statusCode) else { return }
- guard let data = data else { return }
- do {
- let userResult = try JSONDecoder().decode(ProductList.self, from: data)
- 
- for index in 0..<userResult.results.count {
- HomeInfoDatas.names.append(userResult.results[index].name)
- HomeInfoDatas.price.append(userResult.results[index].price)
- HomeInfoDatas.images.append(userResult.results[index].main_img[0].image)
- HomeInfoDatas.brandNames.append(userResult.results[index].brand.name)
- 
- for imageIndex in 0..<userResult.results[index].main_img.count {
- if !(HomeInfoDatas.idAndImages.keys.contains(userResult.results[index].name)) {
- HomeInfoDatas.idAndImages[userResult.results[index].name] = [imageIndex : userResult.results[index].main_img[imageIndex].image]
- } else {
- HomeInfoDatas.idAndImages[userResult.results[index].name]?.updateValue(userResult.results[index].main_img[imageIndex].image, forKey: imageIndex)
- }
- }
- self.firstToggle = true
- }
- } catch {
- print ("failed to convert error : ", error.localizedDescription)
- }
- }
- task.resume()
- }
- 
- private func getProductDetailImage() {
- let productUrl = "http://52.78.75.94/events/"
- guard let url = URL(string: productUrl) else { return }
- let task = URLSession.shared.dataTask(with: url) { (data, response, error) in
- guard error == nil else { return print ("error : ", error!.localizedDescription)}
- let responsea = response as? HTTPURLResponse
- print (responsea!.statusCode)
- guard let response = response as? HTTPURLResponse,
- (200..<406).contains(response.statusCode) else { return }
- guard let data = data else { return }
- do {
- let detailImages = try JSONDecoder().decode([Events].self, from: data)
- for index in 0..<detailImages.count {
- HomeInfoDatas.bannerImages.append(detailImages[index].images)
- self.secondToggle = true
- }
- } catch {
- print ("failed to convert error : ", error.localizedDescription)
- }
- }
- task.resume()
- }
- 
- */
